*Finally* identified a Frac repr that allows me to see the features at 50x zoom.
The Frac repr is in `mod bampf` added here; "BAMPF", besides being the
nightcrawler teleport sound, stands for "Big (as in) Arbitrary
Mantissa Precision Floats"

With that in place, I changed some other code (in an attempt to
improve the amount of precision we would maintain).

The code overall is still in desparate need of cleanup, but at least
now I have something I am happy presenting.